fixed graph

+++ src/utils/makeDate.js
@@ -1,9 +1,7 @@
 export default function makeDate(date)
 {
-  return Date.UTC(
-    ...([...date]
-      .flatMap((ele) => (!isNaN(parseInt(ele)) ? parseInt(ele) : " "))
-      .join("")).split(" ")
-  );
+  return Date.UTC(...([...date]
+    .flatMap((ele) => (!isNaN(parseInt(ele)) ? parseFloat(ele) : " "))
+    .join("")).split(" "));
 
 }
